fix: use Decimal type for SportsEvent.volume to match Numeric column

+++ api/data/sports/models.py
@@ -1,5 +1,6 @@
 # api/data/sports/models.py
 from datetime import datetime, UTC
+from decimal import Decimal
 from sqlalchemy import String, DateTime, Numeric, Integer, JSON
 from sqlalchemy.orm import Mapped, mapped_column
 from db.postgres import Base
@@ -15,7 +16,7 @@ class SportsEvent(Base):
     outcomes: Mapped[list] = mapped_column(JSON, nullable=False, default=list)
     status: Mapped[str] = mapped_column(String(20), nullable=False, default="active")
     resolution: Mapped[dict | None] = mapped_column(JSON, nullable=True)
-    volume: Mapped[float | None] = mapped_column(Numeric(20, 6), nullable=True)
+    volume: Mapped[Decimal | None] = mapped_column(Numeric(20, 6), nullable=True)
     end_date: Mapped[datetime | None] = mapped_column(DateTime(timezone=True), nullable=True)
     data_updated_at: Mapped[datetime] = mapped_column(
         DateTime(timezone=True), nullable=False, default=lambda: datetime.now(UTC)
